+++ playground/Optsol.Playground.Api/Controllers/Clientes/ClientesController.cs
@@ -12,30 +12,28 @@ using Optsol.Playground.Application.Services.Cliente;
 using Optsol.Playground.Application.ViewModels.Cliente;
 using Optsol.Playground.Domain.Entities;
 
-namespace Optsol.Playground.Api.Controllers.Clientes
-{
-    [Authorize]
-    [ApiController]
-    [Authorize(AuthenticationSchemes = "Bearer")]
-    [Route("api/[Controller]")]
-    public partial class ClientesController : ApiControllerBase<ClientePessoaFisicaEntity, ClienteRequest, ClienteResponse, ClienteSearchDto>
-    {
-        public readonly IClienteServiceApplication _clienteServiceApplication;
+namespace Optsol.Playground.Api.Controllers.Clientes;
 
-        public ClientesController(
-            ILoggerFactory logger,
-            IResponseFactory responseFactory,
-            IClienteServiceApplication clienteServiceApplication)
-            : base(logger, clienteServiceApplication, responseFactory)
-        {
-            _clienteServiceApplication = clienteServiceApplication;
-            _clienteServiceApplication.Includes = clientes => clientes.Include(x => x.Cartoes);
-        }
-        
-        [OptsolAuthorize("ClaimTeste", "ClaimTeste2")]
-        public override Task<IActionResult> GetByIdAsync(Guid id)
-        {
-            return base.GetByIdAsync(id);
-        }
+[ApiController]
+[Authorize(AuthenticationSchemes = "Bearer")]
+[Route("api/[Controller]")]
+public partial class ClientesController : ApiControllerBase<ClientePessoaFisicaEntity, ClienteRequest, ClienteResponse, ClienteSearchDto>
+{
+    private readonly IClienteServiceApplication clienteServiceApplication;
+
+    public ClientesController(
+        ILoggerFactory logger,
+        IResponseFactory responseFactory,
+        IClienteServiceApplication clienteServiceApplication)
+        : base(logger, clienteServiceApplication, responseFactory)
+    {
+        this.clienteServiceApplication = clienteServiceApplication;
+        this.clienteServiceApplication.Includes = clientes => clientes.Include(x => x.Cartoes);
+    }
+    
+    [OptsolAuthorize("ClaimTeste", "ClaimTeste2")]
+    public override Task<IActionResult> GetByIdAsync(Guid id)
+    {
+        return base.GetByIdAsync(id);
     }
 }
