+++ playground/Optsol.Playground.Api/appsettings.Development.json
@@ -17,7 +17,7 @@
     "Authority": {
       "ClientId": "40b37822-1e85-4977-8a9b-fdc0e47ed2d4",
       "Instance": "https://optsolssodev.b2clogin.com",
-      "Domain": "optsolssodev.onmicrosoft.com"  
+      "Domain": "optsolssodev.onmicrosoft.com"
     }
   },
   "FirebaseSettings": {
