+++ playground/Optsol.Playground.Api/appsettings.json
@@ -1,4 +1,7 @@
 {
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=127.0.0.1,1433;Database=Playground;User=sa;Password=OPTSOL@dev;MultipleActiveResultSets=true"
+  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
