+++ src/Optsol.Components.Infra.Bus/Optsol.Components.Infra.Bus.csproj
@@ -4,7 +4,7 @@
         <TargetFramework>net6.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-             <Version>2.3.0</Version>
+        <Version>2.3.3</Version>
     </PropertyGroup>
 
 </Project>
