+++ src/Optsol.Components.Infra.ElasticSearch/Optsol.Components.Infra.ElasticSearch.csproj
@@ -1,10 +1,10 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
-    <Version>2.3.0</Version>
+    <Version>2.3.3</Version>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="NEST" Version="7.17.0" />
+    <PackageReference Include="NEST" Version="7.17.5" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Optsol.Components.Infra\Optsol.Components.Infra.csproj" />
