+++ src/Optsol.Components.Infra.Security.AzureB2C/Optsol.Components.Infra.Security.AzureB2C.csproj
@@ -4,7 +4,7 @@
         <TargetFramework>net6.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <Version>2.3.0</Version>
+        <Version>2.3.3</Version>
     </PropertyGroup>
     
     <ItemGroup>
