+++ src/Optsol.Components.Infra/Data/Provider/ITenantProvider.cs
@@ -4,9 +4,10 @@ namespace Optsol.Components.Infra.Data.Provider
 {
     public interface ITenantProvider<out TKey>
     {
-        TKey GetTenantId();
+        TKey TenantId { get; }
     }
 
+    /// <inheritdoc />
     public interface ITenantProvider : ITenantProvider<Guid>
     {
