+++ src/Optsol.Components.Infra/Data/Repository.cs
@@ -18,39 +18,40 @@ namespace Optsol.Components.Infra.Data
     public class Repository<TEntity, TKey> : IRepository<TEntity, TKey>
         where TEntity : class, IAggregateRoot<TKey>
     {
-        private bool _disposed = false;
+        private bool disposed;
 
-        private readonly ILogger _logger;
+        private readonly ILogger logger;
 
-        private readonly ITenantProvider<TKey> _tenantProvider;
+        private readonly ITenantProvider<TKey> tenantProvider;
 
+        // ReSharper disable once MemberCanBePrivate.Global
         public CoreContext Context { get; protected set; }
 
+        // ReSharper disable once MemberCanBeProtected.Global
         public DbSet<TEntity> Set { get; protected set; }
 
         public Repository(CoreContext context, ILoggerFactory logger, ITenantProvider<TKey> tenantProvider = null)
         {
-            _logger = logger.CreateLogger(nameof(Repository<TEntity, TKey>));
-            _logger?.LogInformation($"Inicializando Repository<{ typeof(TEntity).Name }, { typeof(TKey).Name }>");
+            this.logger = logger.CreateLogger(nameof(Repository<TEntity, TKey>));
+            this.logger?.LogInformation($"Inicializando Repository<{ typeof(TEntity).Name }, { typeof(TKey).Name }>");
 
             Context = context ?? throw new DbContextNullException();
-            this.Set = context.Set<TEntity>();
+            Set = context.Set<TEntity>();
 
-            _tenantProvider = tenantProvider;
-            ValidateTentatProvider();
+            this.tenantProvider = tenantProvider;
+            ValidateTenantProvider();
         }
 
-        private void ValidateTentatProvider()
+        private void ValidateTenantProvider()
         {
             TypeFilter filter = new(InterfaceFilter);
 
             var @interface = $"{typeof(ITenant<TKey>).Namespace}.{typeof(ITenant<TKey>).Name.Replace("`1", "")}";
-            var repositoryInvalid = typeof(TEntity).FindInterfaces(filter, @interface).Any() && _tenantProvider == null;
-            if (repositoryInvalid)
-            {
-                _logger?.LogError($"Essa entidade implementa ITenant, o ITenantProvider deve ser injetado.");
-                throw new InvalidRepositoryException();
-            }
+            var repositoryInvalid = typeof(TEntity).FindInterfaces(filter, @interface).Any() && tenantProvider == null;
+            if (!repositoryInvalid) return;
+            
+            logger?.LogError($"Essa entidade implementa ITenant, o ITenantProvider deve ser injetado.");
+            throw new InvalidRepositoryException();
         }
 
         private static bool InterfaceFilter(Type typeObj, Object criteriaObj)
@@ -63,21 +64,21 @@ namespace Optsol.Components.Infra.Data
 
         public virtual Task<TEntity> GetByIdAsync(TKey id)
         {
-            _logger?.LogInformation($"Método: { nameof(GetByIdAsync) }( {{ id:{ id } }} ) Retorno: type { typeof(TEntity).Name }");
+            logger?.LogInformation($"Método: { nameof(GetByIdAsync) }( {{ id:{ id } }} ) Retorno: type { typeof(TEntity).Name }");
 
             return Set.FindAsync(id).AsTask();
         }
 
         public virtual Task<IEnumerable<TEntity>> GetByIdsAsync(IEnumerable<TKey> ids)
         {
-            _logger?.LogInformation($"Método: { nameof(GetByIdAsync) }( {{ ids:[{ string.Join(",", ids) }]}} ) Retorno: type { typeof(TEntity).Name }");
+            logger?.LogInformation($"Método: { nameof(GetByIdAsync) }( {{ ids:[{ string.Join(",", ids) }]}} ) Retorno: type { typeof(TEntity).Name }");
 
             return Set.Where(a => ids.Contains(a.Id)).AsAsyncEnumerable().AsyncEnumerableToEnumerable();
         }
 
         public virtual Task<IEnumerable<TEntity>> GetByIdsAsync(IEnumerable<TKey> ids, Func<IQueryable<TEntity>, IQueryable<TEntity>> includes)
         {
-            _logger?.LogInformation($"Método: { nameof(GetByIdAsync) }( {{ ids:[{ string.Join(",", ids) }]}} ) Retorno: type { typeof(TEntity).Name }");
+            logger?.LogInformation($"Método: { nameof(GetByIdAsync) }( {{ ids:[{ string.Join(",", ids) }]}} ) Retorno: type { typeof(TEntity).Name }");
 
             var querable = Set.AsQueryable();
 
@@ -92,7 +93,7 @@ namespace Optsol.Components.Infra.Data
 
         public virtual Task<TEntity> GetByIdAsync(TKey id, Func<IQueryable<TEntity>, IQueryable<TEntity>> Includes)
         {
-            _logger?.LogInformation($"Método: { nameof(GetByIdAsync) }( {{ id:{ id } }} ) Retorno: type { typeof(TEntity).Name }");
+            logger?.LogInformation($"Método: { nameof(GetByIdAsync) }( {{ id:{ id } }} ) Retorno: type { typeof(TEntity).Name }");
 
             var querable = Set.AsQueryable();
 
@@ -107,14 +108,14 @@ namespace Optsol.Components.Infra.Data
 
         public virtual Task<IEnumerable<TEntity>> GetAllAsync()
         {
-            _logger?.LogInformation($"Método: { nameof(GetAllAsync) }() Retorno: IAsyncEnumerable<{ typeof(TEntity).Name }>");
+            logger?.LogInformation($"Método: { nameof(GetAllAsync) }() Retorno: IAsyncEnumerable<{ typeof(TEntity).Name }>");
 
             return Set.AsAsyncEnumerable().AsyncEnumerableToEnumerable();
         }
 
         public virtual Task<IEnumerable<TEntity>> GetAllAsync(Func<IQueryable<TEntity>, IQueryable<TEntity>> Includes)
         {
-            _logger?.LogInformation($"Método: { nameof(GetAllAsync) }() Retorno: IAsyncEnumerable<{ typeof(TEntity).Name }>");
+            logger?.LogInformation($"Método: { nameof(GetAllAsync) }() Retorno: IAsyncEnumerable<{ typeof(TEntity).Name }>");
 
             var querable = Set.AsQueryable();
 
@@ -147,13 +148,13 @@ namespace Optsol.Components.Infra.Data
 
         public virtual Task InsertAsync(TEntity entity)
         {
-            _logger?.LogInformation($"Método: { nameof(InsertAsync) }( {{entity:{ entity.ToJson() }}} )");
+            logger?.LogInformation($"Método: { nameof(InsertAsync) }( {{entity:{ entity.ToJson() }}} )");
 
             var entityIsITenant = entity is ITenant<TKey>;
             if (entityIsITenant)
             {
-                _logger?.LogInformation($"Executando SetTenantId({_tenantProvider.GetTenantId()}) em InsertAsync");
-                ((ITenant<TKey>)entity).SetTenantId(_tenantProvider.GetTenantId());
+                logger?.LogInformation($"Executando SetTenantId({tenantProvider.TenantId}) em InsertAsync");
+                ((ITenant<TKey>)entity).SetTenantId(tenantProvider.TenantId);
             }
 
             return Set.AddAsync(entity).AsTask();
@@ -161,7 +162,7 @@ namespace Optsol.Components.Infra.Data
 
         public virtual Task UpdateAsync(TEntity entity)
         {
-            _logger?.LogInformation($"Método: { nameof(UpdateAsync) }( {{entity:{ entity.ToJson() }}} )");
+            logger?.LogInformation($"Método: { nameof(UpdateAsync) }( {{entity:{ entity.ToJson() }}} )");
 
             SetDetachedLocalEntity(entity);
 
@@ -183,11 +184,11 @@ namespace Optsol.Components.Infra.Data
 
         private void SetTenantIdFromTenantProvider(TEntity entity)
         {
-            if (entity is not ITenant<TKey>)
+            if (entity is not ITenant<TKey> tenant)
                 return;
 
-            _logger?.LogInformation($"Executando SetTenantId({_tenantProvider.GetTenantId()}) em UpdateAsync");
-            ((ITenant<TKey>)entity).SetTenantId(_tenantProvider.GetTenantId());
+            logger?.LogInformation($"Executando SetTenantId({tenantProvider.TenantId}) em UpdateAsync");
+            tenant.SetTenantId(tenantProvider.TenantId);
 
         }
 
@@ -198,7 +199,7 @@ namespace Optsol.Components.Infra.Data
             var entityNotFound = entity == null;
             if (entityNotFound)
             {
-                _logger?.LogError($"Método: { nameof(DeleteAsync) }({{ TKey:{ id.ToJson() } }}) Registro não encontrado");
+                logger?.LogError($"Método: { nameof(DeleteAsync) }({{ TKey:{ id.ToJson() } }}) Registro não encontrado");
                 return;
             }
 
@@ -207,7 +208,7 @@ namespace Optsol.Components.Infra.Data
 
         public virtual async Task DeleteAsync(TEntity entity)
         {
-            _logger?.LogInformation($"Método: { nameof(DeleteAsync) }( {{entity:{ entity.ToJson() }}} )");
+            logger?.LogInformation($"Método: { nameof(DeleteAsync) }( {{entity:{ entity.ToJson() }}} )");
 
             var entityFound = entity != null;
             var entityIsDeletable = entity is IDeletable;
@@ -219,13 +220,13 @@ namespace Optsol.Components.Infra.Data
             }
             else
             {
-                Set.Attach(entity).State = EntityState.Deleted;
+                if (entity != null) Set.Attach(entity).State = EntityState.Deleted;
             }
         }
 
         public virtual Task<int> SaveChangesAsync()
         {
-            _logger?.LogInformation($"Método: { nameof(SaveChangesAsync) }()");
+            logger?.LogInformation($"Método: { nameof(SaveChangesAsync) }()");
 
             return Context.SaveChangesAsync();
         }
@@ -238,13 +239,13 @@ namespace Optsol.Components.Infra.Data
 
         protected virtual void Dispose(bool disposing)
         {
-            _logger?.LogInformation($"Método: { nameof(Dispose) }()");
+            logger?.LogInformation($"Método: { nameof(Dispose) }()");
 
-            if (!_disposed && disposing)
+            if (!disposed && disposing)
             {
                 Context.Dispose();
             }
-            _disposed = true;
+            disposed = true;
         }
 
         private static async Task<ISearchResult<TEntity>> CreateSearchResult(IQueryable<TEntity> query, uint page, uint? pageSize)
@@ -262,7 +263,8 @@ namespace Optsol.Components.Infra.Data
             return searchResult;
         }
 
-        private static IQueryable<TEntity> ApplyPagination(IQueryable<TEntity> query, uint page, uint? pageSize)
+        // ReSharper disable once MemberCanBePrivate.Global
+        protected static IQueryable<TEntity> ApplyPagination(IQueryable<TEntity> query, uint page, uint? pageSize)
         {
             var skip = --page * (pageSize ?? 0);
