+++ src/Optsol.Components.Infra/InfraConstants.cs
@@ -1,0 +1,6 @@
+namespace Microsoft.Extensions.DependencyInjection;
+
+public static class InfraConstants
+{
+    public const string TenantId = "6025384C-3CF7-4310-AFA9-244507D1FE9B";
+}
