+++ test/Optsol.Components.Test.Unit/Infra/RepositoryOptionsSpec.cs
@@ -1,0 +1,12 @@
+using Xunit;
+
+namespace Optsol.Components.Test.Unit.Infra;
+
+public class RepositoryOptionsSpec
+{
+    [Trait("RepositoryOptions", "Adicionando TenantInterceptor")]
+    public void DeveAdicionarTenantInterceptor()
+    {
+        
+    }
+}
