+++ test/Optsol.Components.Test.Utils/Data/Repositories/Tenants/ITestTenantReadRepository.cs
@@ -1,8 +1,8 @@
+using System;
 using Optsol.Components.Domain.Data;
 using Optsol.Components.Test.Utils.Data.Entities;
-using System;
 
-namespace Optsol.Components.Test.Utils.Repositories.Tenant
+namespace Optsol.Components.Test.Utils.Data.Repositories.Tenants
 {
     public interface ITestTenantReadRepository: IReadRepository<TestTenantEntity, Guid>
     {
