+++ test/Optsol.Components.Test.Utils/Data/Repositories/Tenants/ITestWriteTenantRepository.cs
@@ -1,8 +1,8 @@
-﻿using Optsol.Components.Domain.Data;
+﻿using System;
+using Optsol.Components.Domain.Data;
 using Optsol.Components.Test.Utils.Data.Entities;
-using System;
 
-namespace Optsol.Components.Test.Utils.Repositories.Tenant
+namespace Optsol.Components.Test.Utils.Data.Repositories.Tenants
 {
     public interface ITestTenantWriteRepository : IWriteRepository<TestTenantEntity, Guid>
     {
