+++ test/Optsol.Components.Test.Utils/Data/Repositories/Tenants/TestTenantReadRepository.cs
@@ -1,23 +1,23 @@
+using System;
 using Microsoft.Extensions.Logging;
 using Optsol.Components.Infra.Data;
 using Optsol.Components.Infra.Data.Provider;
 using Optsol.Components.Test.Utils.Data.Entities;
-using System;
 
-namespace Optsol.Components.Test.Utils.Repositories.Tenant
+namespace Optsol.Components.Test.Utils.Data.Repositories.Tenants
 {
     public class TestTenantReadRepository : Repository<TestTenantEntity, Guid>, ITestTenantReadRepository
     {
-        public TestTenantReadRepository(CoreContext context, ILoggerFactory logger, ITenantProvider tenantProvider) 
-            : base(context, logger, tenantProvider)
+        public TestTenantReadRepository(CoreContext context, ILoggerFactory logger, ITenantProvider loggedUser) 
+            : base(context, logger, loggedUser)
         {
         }
     }
 
     public class TestTenantDeletableReadRepository : Repository<TestTenantDeletableEntity, Guid>, ITestTenantDeletableReadRepository
     {
-        public TestTenantDeletableReadRepository(CoreContext context, ILoggerFactory logger, ITenantProvider tenantProvider)
-            : base(context, logger, tenantProvider)
+        public TestTenantDeletableReadRepository(CoreContext context, ILoggerFactory logger, ITenantProvider loggedUser)
+            : base(context, logger, loggedUser)
         {
         }
     }
